+++ src/pc/index.js
@@ -6,27 +6,36 @@
 const prompt = require('prompt-sync')()
 const chalk = require('chalk')
 
-const zaDarkPC = require('./zadark-pc')
-const { log, logError, open, isRoot } = require('./utils')
-const { PLATFORM, ZADARK_VERSION } = require('./constants')
+const zadarkPC = require('./zadark-pc')
+const { log, logError, open, isRoot, killProcess } = require('./utils')
+
+const { ZADARK_VERSION, IS_MAC } = require('./constants')
 
 const renderHeader = () => {
   log('')
-  log(chalk.blueBright.bold('ZaDark – Zalo Dark Mode'))
-  log(chalk.blueBright('Website   :', chalk.underline('https://zadark.quaric.com')))
-  log(chalk.blueBright('Phien ban :', `${PLATFORM === 'darwin' ? 'macOS' : 'Windows'}-${ZADARK_VERSION}`))
+  log(chalk.blueBright.bold(`ZaDark for ${IS_MAC ? 'macOS' : 'Windows'} ${ZADARK_VERSION}`))
+  log(chalk.blueBright(chalk.underline('https://zadark.quaric.com')))
   log('')
-  log(chalk.blueBright('<ZaDark from Quaric>'))
-  log('')
+}
+
+const quitZalo = async () => {
+  const zaloPID = await zadarkPC.getZaloProcessId()
+  if (zaloPID) {
+    log('')
+    log(chalk('>> Dang thoat Zalo PC ...'))
+    await killProcess(zaloPID)
+  }
 }
 
 const handleInstall = async (zaloResDirList) => {
   log(chalk.magentaBright.bold('[CAI DAT ZADARK]'))
 
+  await quitZalo()
+
   for (const zaloResDir of zaloResDirList) {
     log('')
     log(chalk('>> Dang cai dat', chalk.bold(zaloResDir)))
-    await zaDarkPC.installDarkTheme(zaloResDir)
+    await zadarkPC.installDarkTheme(zaloResDir)
   }
 
   log('')
@@ -34,12 +43,14 @@ const handleInstall = async (zaloResDirList) => {
 }
 
 const handleUninstall = async (zaloResDirList) => {
-  log(chalk.magentaBright.bold('[GOI CAI DAT ZADARK]'))
+  log(chalk.magentaBright.bold('[GO CAI DAT ZADARK]'))
+
+  await quitZalo()
 
   for (const zaloResDir of zaloResDirList) {
     log('')
     log(chalk('>> Dang go cai dat', chalk.bold(zaloResDir)))
-    await zaDarkPC.uninstallDarkTheme(zaloResDir)
+    await zadarkPC.uninstallDarkTheme(zaloResDir)
   }
 
   log('')
@@ -47,18 +58,21 @@ const handleUninstall = async (zaloResDirList) => {
 }
 
 (async () => {
+  let featureIndex
+
   try {
-    if (PLATFORM === 'darwin' && !isRoot()) {
+    if (IS_MAC && !isRoot()) {
       const supportUrl = 'https://zadark.quaric.com/pc/macos#run-zadark-as-root'
       open(supportUrl)
       throw new Error(`Vui long chay ZaDark voi quyen Root (${supportUrl}).`)
     }
 
-    const defaultZaloPath = PLATFORM === 'darwin'
+    const defaultZaloPath = IS_MAC
       ? '/Applications/Zalo.app' // ! Can't be changed
       : process.env.USERPROFILE + '/AppData/Local/Programs/Zalo' // ! Can't be changed
 
-    let zaloResDirList = zaDarkPC.getZaloResDirList(defaultZaloPath)
+    let zaloResDirList = zadarkPC.getZaloResDirList(defaultZaloPath)
+
     if (!zaloResDirList.length) {
       log('')
       log(chalk.redBright(`Khong tim thay Zalo PC tai duong dan mac dinh (${defaultZaloPath}).`))
@@ -67,10 +81,11 @@ const handleUninstall = async (zaloResDirList) => {
       log(chalk.magentaBright.bold('[TUY CHINH DUONG DAN ZALO]'))
       log('')
 
-      const examplePath = PLATFORM === 'darwin' ? '/ThuMucDaCaiZalo/Zalo.app' : 'D:\\ThuMucDaCaiZalo\\Zalo'
+      const examplePath = IS_MAC ? '/ThuMucDaCaiZalo/Zalo.app' : 'D:\\ThuMucDaCaiZalo\\Zalo'
       const customZaloPath = prompt(chalk.yellowBright(`> Nhap duong dan Zalo PC cua ban (VD: ${examplePath}) : `))
 
-      zaloResDirList = zaDarkPC.getZaloResDirList(customZaloPath)
+      zaloResDirList = zadarkPC.getZaloResDirList(customZaloPath)
+
       if (!zaloResDirList.length) {
         throw new Error('Khong tim thay Zalo PC. Vui long tai va cai dat Zalo PC tai "https://zalo.me/pc".')
       }
@@ -79,18 +94,10 @@ const handleUninstall = async (zaloResDirList) => {
     console.clear()
     renderHeader()
 
-    log(chalk.magentaBright.bold('[LUU Y QUAN TRONG]'))
+    log(chalk.redBright('- DE DAM BAO AN TOAN, vui long tai ZaDark tai:'))
+    log(' ', chalk.underline.redBright('https://sourceforge.net/projects/zadark/files/zadarkPC'))
+    log(chalk.redBright('- Vui long cai dat lai ZaDark sau khi cap nhat Zalo PC.'))
     log('')
-    log(chalk.redBright('1. DE DAM BAO AN TOAN, truoc khi cai dat hay chac chan ban da tai ZaDark tu'))
-    log('  ', chalk.underline.redBright('https://sourceforge.net/projects/zadark/files/ZaDarkPC/'))
-    log('2. Vui long thoat Zalo PC truoc khi cai dat, go cai dat ZaDark.')
-    log('3. Vui long cai dat lai ZaDark sau khi cap nhat Zalo PC.')
-    log('')
-
-    prompt(chalk.yellowBright('> Nhan', chalk.bold('[enter]'), 'de bat dau ...'))
-
-    console.clear()
-    renderHeader()
 
     log(chalk.magentaBright.bold('[CHUC NANG]'))
     log('')
@@ -99,7 +106,7 @@ const handleUninstall = async (zaloResDirList) => {
     log('3. Thoat')
     log('')
 
-    const featureIndex = prompt(chalk.yellowBright('> Nhap STT chuc nang', chalk.bold('[1-3]'), 'va nhan', chalk.bold('[enter]'), ': '))
+    featureIndex = prompt(chalk.yellowBright('> Nhap STT chuc nang', chalk.bold('[1-3]'), 'va nhan', chalk.bold('[enter]'), ': '))
 
     console.clear()
     renderHeader()
@@ -128,6 +135,9 @@ const handleUninstall = async (zaloResDirList) => {
     log('Cam on ban da su dung ZaDark, chuc ban lam viec hieu qua!')
     log('Goodbye.')
     log('')
-    prompt(chalk.yellowBright('> Nhan', chalk.bold('[enter]'), 'de thoat chuong trinh ...'))
+
+    if (['1', '2'].includes(featureIndex)) {
+      prompt(chalk.yellowBright('> Nhan', chalk.bold('[enter]'), 'de thoat chuong trinh ...'))
+    }
   }
 })()