+++ src/pc/zadark-pc.js
@@ -10,15 +10,16 @@ const asar = require('@electron/asar')
 const HTMLParser = require('node-html-parser')
 const glob = require('glob')
 
-const { logDebug, copyRecursiveSync } = require('./utils')
-const { PLATFORM, ZADARK_VERSION } = require('./constants')
+const { logDebug, copyRecursiveSync, getProcessIdByName } = require('./utils')
+
+const { PLATFORM, ZADARK_VERSION, IS_MAC } = require('./constants')
 
 const getZaloResDirList = (customZaloPath) => {
   if (!['darwin', 'win32'].includes(PLATFORM)) {
     throw new Error(`Khong ho tro he dieu hanh "${PLATFORM}".`)
   }
 
-  const resourcesPath = PLATFORM === 'darwin'
+  const resourcesPath = IS_MAC
     ? path.join(customZaloPath, './Contents/Resources')
     : path.join(customZaloPath, './Zalo-*/resources')
 
@@ -180,7 +181,7 @@ const writeZNotificationFile = (zaloDir) => {
 
   removeZaDarkCSSAndJS({ headElement, bodyElement })
 
-  // Required fonts, stylesheets
+  // Required fonts, stylesheets and scripts
   const elements = [
     {
       selector: 'link[href="zadark-fonts.min.css"]',
@@ -193,6 +194,12 @@ const writeZNotificationFile = (zaloDir) => {
       where: 'beforeend',
       html: '<link rel="stylesheet" href="zadark-znotification.min.css">',
       htmlElement: headElement
+    },
+    {
+      selector: 'script[src="zadark-znotification.min.js"]',
+      where: 'beforeend',
+      html: '<script src="zadark-znotification.min.js"></script>',
+      htmlElement: bodyElement
     }
   ]
   elements.forEach((element) => {
@@ -288,14 +295,16 @@ const installDarkTheme = async (zaloDir) => {
     copyAssetDir(zaloDir, asset)
   })
 
-  // Add "themeAttributes, classNames, fonts, stylesheets" to "resources/app/pc-dist/index.html"
+  // Add "themeAttributes, classNames, fonts, stylesheets and scripts" to "resources/app/pc-dist/index.html"
   writeIndexFile(zaloDir)
 
   // Add zadark-main to "resources/app/bootstrap.js"
   writeBootstrapFile(zaloDir)
 
-  // Add fonts, stylesheets" to "resources/app/pc-dist/znotification.html"
-  writeZNotificationFile(zaloDir)
+  if (!IS_MAC) {
+    // Add fonts, stylesheets and scripts" to "resources/app/pc-dist/znotification.html"
+    writeZNotificationFile(zaloDir)
+  }
 
   // Create package "resources/app.asar" from "resources/app" -> Delete "resources/app"
   await asar.createPackage(appDirPath, appAsarPath)
@@ -306,16 +315,27 @@ const uninstallDarkTheme = async (zaloDir) => {
   const appAsarPath = path.join(zaloDir, 'app.asar')
   const appAsarBakPath = path.join(zaloDir, 'app.asar.bak')
 
+  // Delete "resources/app.asar"
   // Rename "resources/app.asar.bak" to "resources/app.asar"
   if (fs.existsSync(appAsarBakPath)) {
+    logDebug('- deleteFile', appAsarPath)
+    await del(appAsarPath, { force: true })
     fs.renameSync(appAsarBakPath, appAsarPath)
     logDebug('- renameFile', appAsarBakPath)
   }
+}
+
+const getZaloProcessId = async () => {
+  const zaloProcessName = IS_MAC ? 'Zalo' : 'Zalo.exe'
+  const zaloProcessId = await getProcessIdByName(zaloProcessName)
+  return zaloProcessId
 }
 
 module.exports = {
   getZaloResDirList,
 
   installDarkTheme,
-  uninstallDarkTheme
+  uninstallDarkTheme,
+
+  getZaloProcessId
 }